Add GitHub service
This Spring Cloud application gathers data from the GitHub API for a
given user: user profile and latest user events.

This application registers itself with the Service Registry and uses
Spring Cloud Sleuth for tracing. Note that for demo purporses, we're
using an `AlwaysSampler` to send all traces, all the time.

The GitHub API is rate-limited, so adding a personal API token can be
quite useful. We've got an existing `github.token` configuration
property for that but no value is set for now.