Add profile web page
This commit creates a profile web page containing information on
speakers coming from both the database and the GitHub information
provided by the github-service.

To retrieve information from the github-service, we need to:
* use a `LoadBalancerClient` to find a service instance
* use a `RestTemplate` to query that remote service

Obviously, this is the "manual" way of doing things, as Spring Cloud can
instrument a `RestTemplate` bean if it's marked as `@LoadBalanced`.