Add tracing to the speaker-service
This commit adds `spring-cloud-starter-zipkin` to the speaker service so
as to trace requests and send them to the zipkin collector. This allows
us to understand how or system is behaving and where latency happens.

Again, this example shows how to manually instrument a `RestTemplate`.
This is done for you for `RestTemplate` beans and will be done as well
for `RestTemplateBuilder` in a future version.

Because this is a demo application, we're also using an `AlwaysSampler`
to make sure we get all traces during the demo.