Get token configuration property from config server
The GitHub API is rate-limited, so we'd like to use a GitHub token to
upgrade the number of requests we're allowed to send per hour.

Since this token is sensitive data, we don't want to commit that as
plaintext in our application configuration.

With this commit, we use the existing config server to fetch our
configuration. The token information is encrypted in the configuration
repository and deciphered on the fly by the config server (which holds a
cipher key).

With this setup, we assume that our config server is located in a DMZ
and that clients are authorized to see that data. Note that there are
other types of setups where clients can decipher configuration data
themselves or use advanced features like Vault.